+++ NEWS.md
@@ -7,6 +7,8 @@
   non-blank lines read from config files.
 - tarsnap now gives an error if there are unused command-line arguments.
   (i.e. "tarsnap -d -f a1 a2", where "a2" is unused.)
+- When sent SIGINFO or SIGUSR1, tarsnap now prints the number of files and the
+  number of uncompressed bytes processed, in addition to the previous output.
 
 
 Tarsnap Releases